Cleanup src tree (#288)
* Move CompilationTaskContext to toolchain, as it doesn't really belong in utils.

* Move KotlinCompilerPluginArgsEncoder to toolchain. Temporary placement, new location TBD

* Isolate the utils package

* Move toolchain into a separate compilation unit. Causes some issues with inlining.

* Split apart test packages.

* Split apart test packages.

* Split apart test packages.

* add bazel version for development

* factor out JvmCompilationTask extensions to reuse in abi generation

* move compiler to source package. add .bazelversion for development
