+++ .bazelproject
@@ -26,8 +26,10 @@ targets:
   //examples/dagger/...
   # These targets are built for the ide only. Primary purpose is to ensure the builder can build the targets, but it's
   # also a good way of testing the intellij plugin.
-  //src/main/kotlin/io/bazel/kotlin/builder:builder_kt_for_ide
-  //src/main/kotlin:compiler_lib_for_ide
+  //src/main/kotlin/io/bazel/kotlin/builder/tasks:tasks_for_ide
+  //src/main/kotlin/io/bazel/kotlin/builder/utils:utils_for_ide
+  //src/main/kotlin/io/bazel/kotlin/builder/toolchain:toolchain_for_ide
+  //src/main/kotlin/io/bazel/kotlin/compiler:compiler_for_ide
 
 test_sources:
   src/test/*
