+++ src/main/kotlin/io/bazel/kotlin/builder/tasks/BUILD.bazel
@@ -1,4 +1,4 @@
-# Copyright 2018 The Bazel Authors. All rights reserved.
+# Copyright 2020 The Bazel Authors. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,26 +11,24 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-load("@rules_java//java:defs.bzl", "java_library")
 
-java_library(
-    name = "builder",
-    srcs = glob(["*.java"]),
+load("//src/main/kotlin:bootstrap.bzl", "kt_bootstrap_library")
+
+kt_bootstrap_library(
+    name = "tasks",
+    srcs = glob([
+        "**/*.kt",
+    ]),
     visibility = ["//src:__subpackages__"],
-    runtime_deps = [
-        "@com_github_jetbrains_kotlin//:kotlin-stdlib-jdk7",
-        "@com_github_jetbrains_kotlin//:kotlin-stdlib-jdk8",
-    ],
     deps = [
         "//src/main/protobuf:deps_java_proto",
         "//src/main/protobuf:kotlin_model_java_proto",
         "//src/main/protobuf:worker_protocol_java_proto",
-        "//third_party:dagger",
-        "@com_github_jetbrains_kotlin//:annotations",
-        "@com_github_jetbrains_kotlin//:kotlin-stdlib",
+        "@com_github_jetbrains_kotlin//:kotlin-preloader",
+        "@kotlin_rules_maven//:com_google_protobuf_protobuf_java",
+        "@kotlin_rules_maven//:com_google_protobuf_protobuf_java_util",
         "@kotlin_rules_maven//:javax_inject_javax_inject",
-         "//src/main/kotlin/io/bazel/kotlin/builder/utils:utils",
-         "//src/main/kotlin/io/bazel/kotlin/builder/toolchain:toolchain",
-         "//src/main/kotlin/io/bazel/kotlin/builder/tasks:tasks",
+        "//src/main/kotlin/io/bazel/kotlin/builder/utils:utils",
+        "//src/main/kotlin/io/bazel/kotlin/builder/toolchain:toolchain",
     ],
 )
