+++ src/main/kotlin/io/bazel/kotlin/builder/toolchain/BUILD.bazel
@@ -1,0 +1,30 @@
+# Copyright 2018 The Bazel Authors. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+
+load("//src/main/kotlin:bootstrap.bzl", "kt_bootstrap_library")
+
+kt_bootstrap_library(
+    name = "toolchain",
+    srcs = glob([
+        "*.kt",
+    ]),
+    visibility = ["//src:__subpackages__"],
+    deps = [
+        "//src/main/protobuf:kotlin_model_java_proto",
+        "@com_github_jetbrains_kotlin//:kotlin-preloader",
+        "@kotlin_rules_maven//:com_google_protobuf_protobuf_java",
+        "@kotlin_rules_maven//:com_google_protobuf_protobuf_java_util",
+        "@kotlin_rules_maven//:javax_inject_javax_inject",
+        "//src/main/kotlin/io/bazel/kotlin/builder/utils:utils",
+    ],
+)
