+++ src/test/kotlin/io/bazel/kotlin/builder/utils/BUILD.bazel
@@ -1,0 +1,25 @@
+# Copyright 2020 The Bazel Authors. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+package(default_visibility = ["//src:__subpackages__"])
+
+load("//src/test/kotlin/io/bazel/kotlin:defs.bzl", "kt_rules_test")
+
+kt_rules_test(
+    name = "SourceJarCreatorTest",
+    srcs = ["jars/SourceJarCreatorTest.java"],
+)
+
+test_suite(
+    name = "utils",
+)
